requirements.txt
cycler==0.10.0
distro==1.5.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
six==1.15.0
sklearn==0.0
tabula-py==2.2.0
threadpoolctl==2.1.0
Unidecode==1.2.0
